Fix overflows problems
It solves some overflow's problem:
- conflict between generation and transmission
- generation after the time gap

Every generation is handled in the slot sequent to the slot in which the generation has occurred.
Therefore the function that handles the resource allocation is called always at the beginning of the slot.

SLOT starts -> mainCV2XttiEnds is called which handles the resource allocation of new generated packets that has been generated in the previous slot
THEN mainCV2XttiStarts is called, which handles the transmission of packets for this current slot.

SLOT ends -> mainCV2XtransmissionEnds is called, which handles the update of the KPI