Updates Reselection Counter decrement
Solves a small bug during the decrement of the RC.
With the previous version the RC was decremented every time a BRid was occuring during the current slot. The decrement was done by using hasFirstResourceThisTbeacon.

In the case of aperiodic traffic it could happen that a UE has no packet to transmit, but a BRid assigned. In such case the RC would be decremented anyway.

This behaviour is solved by introducing stationManagement.hasTransmissionThisSlot which is evaluated during mainCV2XttiStarts.

Now the RC is decremented ONLY if:
1) the user transmitted during the slot
2) the transmission was the FIRST transmission

Previously the RC was decremented after the last retransmission. Instead now the decrement is always after the first transmission.
This, is because when RC reaches value 1, and the RC is not updated due to the pkeep, the subsequent transmission occurs without the reservation at future next RRI. The transmission without reservation must occur for both transmission AND retransmission.
Sequently, when RC=0 and a new packet arrives a new resource selection is triggered.

SCHEME:
- RC=2 -> UE1 transmits and RC is decremented to 1.
The test with respect to pkeep fails and the resource is not maintained.
- New packet arrival with RC=1 -> nothing happens
- Transmission with RC=1 -> During the sensing other UEs do not reserve the future resource at timeNow+RRI
- After the sensing, during BRreassignment the RC is decremented to 0
- At the new packet arrival RC=0 and a new resource selection is triggered.